+++ packages/manager/modules/veeam-enterprise/src/dashboard/controller.js
@@ -1,6 +1,6 @@
 import 'moment';
 
-import { RENEW_URL } from '../constants';
+import { RENEW_URL } from '../details/constants';
 
 export default class VeeamEnterpriseDashboardCtrl {
   /* @ngInject */
