+++ packages/manager/modules/veeam-enterprise/src/details/constants.js
