+++ packages/manager/modules/veeam-enterprise/src/details/controller.js
