+++ packages/manager/modules/veeam-enterprise/src/details/index.less
