+++ packages/manager/modules/veeam-enterprise/src/details/service.js
