+++ packages/manager/modules/veeam-enterprise/src/details/template.html
