+++ packages/manager/modules/veeam-enterprise/src/details/translations/Messages_de_DE.json
