+++ packages/manager/modules/veeam-enterprise/src/details/translations/Messages_en_GB.json
