+++ packages/manager/modules/veeam-enterprise/src/details/translations/Messages_fr_CA.json
