+++ packages/manager/modules/veeam-enterprise/src/details/translations/Messages_it_IT.json
