+++ packages/manager/modules/veeam-enterprise/src/details/translations/Messages_lt_LT.json
