+++ packages/manager/modules/veeam-enterprise/src/details/translations/Messages_pt_PT.json
