+++ packages/manager/modules/veeam-enterprise/src/veeam.module.js
