--- packages/manager/modules/veeam-enterprise/src/veeam/translations/Messages_fr_FR.json
@@ -1,3 +1,0 @@
-{
-  "veeam_enterprise_title": "Veeam Enterprise"
-}
